function_modules/sprocket-filter/requirements.txt
ffmpeg-python
redis
pilgram
zlib-state
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
Pillow==6.2.2